DBSync : mounts_templates, emotes Add characters UseHarnessColor (tinyint), mounts_templates FamilyId (int) Harness done

+++ Server/Stump.Server.WorldServer/Game/Conditions/Criterions/AlignementLevelCriterion.cs
@@ -15,7 +15,7 @@ namespace Stump.Server.WorldServer.Game.Conditions.Criterions
 
         public override bool Eval(Character character)
         {
-            return Compare(character.AlignmentGrade, Level);
+            return Compare((int)character.AlignmentGrade, Level);
         }
 
         public override void Build()
